+++ src/interfaces/cache.interface.ts
@@ -1,4 +1,14 @@
-export interface ICache {
-  set: (key: string, value: any) => void;
-  get: (key: string) => any;
+import { IConfig } from "./config.interface";
+import { IFile } from "./dir-scanner.interface";
+
+export interface ISnippet {
+  configMode: "prod" | "dev" | "all";
+  value: string;
+}
+
+export interface IBuilderCache {
+  config: IConfig;
+  templates: { [key: string]: IFile };
+  posts: IFile[];
+  snippets: { [key: string]: ISnippet };
 }