+++ test/unit/builder.css.test.ts
@@ -1,13 +1,25 @@
 jest.mock("uglifycss");
 import { processString } from "uglifycss";
 
-import { MemoryCache } from "../../src/lib/cache";
-import { IFile } from "../../src/interfaces";
+import { MemoryCache, ICache } from "../../src/lib/cache";
+import { IFile, IBuilderCache } from "../../src/interfaces";
 import { buildCss } from "../../src/domain/builder";
+import { defaultConfig } from "../../src/use-cases/build/config.default";
 
 describe("Builder - CSS", () => {
+  let cache: ICache<IBuilderCache>;
+
+  beforeEach(() => {
+    const initCache: IBuilderCache = {
+      config: defaultConfig,
+      templates: {},
+      posts: [],
+      snippets: {}
+    };
+    cache = new MemoryCache<IBuilderCache>(initCache);
+  });
+
   test("Uglify css", async () => {
-    const cache = new MemoryCache();
     const fakeFile: IFile = {
       name: "style",
       extension: "css",