+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated node to 12.13 LTS.
 - File metadata name doesn't include extension.
 - Templates are stored as a IFile.
+- MemoryCache improved typings.
 
 ## Fixed